Create render text class
