+++ src/main.cpp
@@ -14,10 +14,8 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
 
-#include <ft2build.h>
-#include FT_FREETYPE_H
-
 // Others
+#include "render/RenderText.h"
 #include "shader.h"
 
 
@@ -32,23 +30,11 @@ struct
     double y;
 } cursorPos;
 
-struct FontChar
-{
-    unsigned int TextureID;
-    glm::fvec2 Size;
-    glm::fvec2 Bearing;
-    unsigned int Advance;
-};
-
 const unsigned int SCR_WIDTH = 800;
 const unsigned int SCR_HEIGHT = 600;
 
 float obj[] = {-0.5f, -0.5f, 0.0f, 0.5f, -0.5f, 0.0f, 0.0f, 0.5f, 0.0f};
 unsigned int indices[] = {0, 1, 2};
-
-// Fonts variables
-std::map<char, FontChar> FontChars;
-unsigned int fontVAO, fontVBO;
 
 
 int main()
@@ -81,76 +67,8 @@ int main()
     }
 
     /* Shader: font */
-    Shader fontShader("glsl/font.vert", "glsl/font.frag");
     glm::mat4 projection = glm::ortho(0.0f, static_cast<float>(SCR_WIDTH), 0.0f, static_cast<float>(SCR_HEIGHT));
-    fontShader.use();
-    glUniformMatrix4fv(glGetUniformLocation(fontShader.getID(), "projection"), 1, GL_FALSE, glm::value_ptr(projection));
-
-    /* Freetype loaded */
-    std::string fontName = std::filesystem::path("data/fonts/arial.ttf").string();
-    FT_Library ft;
-    FT_Face face;
-
-    if (fontName.empty())
-    {
-        std::cout << "ERROR::FREETYPE: Failed to load font" << std::endl;
-    }
-
-    if (FT_Init_FreeType(&ft))
-    {
-        std::cout << "ERROR::FREETYPE: Could not init FreeType Library" << std::endl;
-        return -1;
-    }
-
-    if (FT_New_Face(ft, fontName.c_str(), 0, &face))
-    {
-        std::cout << "ERROR::FREETYPE: Failed to load font" << std::endl;
-        return -1;
-    }
-    else
-    {
-        FT_Set_Pixel_Sizes(face, 0, 24);
-        glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
-
-        for (unsigned char ch = 0; ch < 128; ch++)
-        {
-            if (FT_Load_Char(face, ch, FT_LOAD_RENDER))
-            {
-                std::cout << "ERROR::FREETYPE: Failed to load Glyph" << std::endl;
-                continue;
-            }
-
-            unsigned int texture;
-            glGenTextures(1, &texture);
-            glBindTexture(GL_TEXTURE_2D, texture);
-            glTexImage2D(GL_TEXTURE_2D, 0, GL_RED, face->glyph->bitmap.width, face->glyph->bitmap.rows, 0, GL_RED,
-                         GL_UNSIGNED_BYTE, face->glyph->bitmap.buffer);
-
-            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-
-            FontChar character = {texture, glm::fvec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
-                                  glm::fvec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                                  static_cast<unsigned int>(face->glyph->advance.x)};
-
-            FontChars.insert(std::pair<char, FontChar>(ch, character));
-        }
-    }
-
-    FT_Done_Face(face);
-    FT_Done_FreeType(ft);
-
-    glGenVertexArrays(1, &fontVAO);
-    glGenBuffers(1, &fontVBO);
-    glBindVertexArray(fontVAO);
-    glBindBuffer(GL_ARRAY_BUFFER, fontVBO);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(float) * 6 * 4, nullptr, GL_DYNAMIC_DRAW);
-    glEnableVertexAttribArray(0);
-    glVertexAttribPointer(0, 4, GL_FLOAT, GL_FALSE, 4 * sizeof(float), nullptr);
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-    glBindVertexArray(0);
+    RenderText rendertext("glsl/font.vert", "glsl/font.frag", projection);
 
     /* Shader: Object */
     Shader objShader("glsl/obj.vert", "glsl/obj.frag");
@@ -196,8 +114,8 @@ int main()
 #ifdef FONT_TEST
         std::stringstream tString;
         tString << "x: " << cursorPos.x << ", y: " << cursorPos.y;
-        render_text(fontShader, tString.str(), static_cast<float>(cursorPos.x),
-                    static_cast<float>(SCR_HEIGHT - cursorPos.y), 1.0f, glm::vec3(0.3f, 0.8f, 0.2f));
+        rendertext.render_text(tString.str(), static_cast<float>(cursorPos.x),
+                               static_cast<float>(SCR_HEIGHT - cursorPos.y), 1.0f, glm::vec3(0.3f, 0.8f, 0.2f));
 #endif
 
         glfwSwapBuffers(window);
@@ -234,51 +152,4 @@ void processInput(GLFWwindow *window)
     {
         glfwSetWindowShouldClose(window, true);
     }
-}
-
-void render_text(Shader &shader, std::string text, float x, float y, float scale, glm::vec3 color)
-{
-    glEnable(GL_CULL_FACE);
-    glEnable(GL_BLEND);
-    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-    glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
-
-    shader.use();
-    glUniform3f(glGetUniformLocation(shader.getID(), "textColor"), color.x, color.y, color.z);
-    glActiveTexture(GL_TEXTURE0);
-    glBindVertexArray(fontVAO);
-
-    std::string::const_iterator iter;
-
-    for (iter = text.begin(); iter != text.end(); iter++)
-    {
-        FontChar ch = FontChars[*iter];
-
-        float xpos = x + ch.Bearing.x * scale;
-        float ypos = y - (ch.Size.y - ch.Bearing.y) * scale;
-        float w = ch.Size.x * scale;
-        float h = ch.Size.y * scale;
-
-        // Update VBO for each character
-        float vertices[6][4] = {{xpos, ypos + h, 0.0f, 0.0f}, {xpos, ypos, 0.0f, 1.0f},
-                                {xpos + w, ypos, 1.0f, 1.0f}, {xpos, ypos + h, 0.0f, 0.0f},
-                                {xpos + w, ypos, 1.0f, 1.0f}, {xpos + w, ypos + h, 1.0f, 0.0f}};
-
-        glBindTexture(GL_TEXTURE_2D, ch.TextureID);
-        glBindBuffer(GL_ARRAY_BUFFER, fontVBO);
-        glBufferSubData(GL_ARRAY_BUFFER, 0, sizeof(vertices), vertices);
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-        // Render Quad
-        glDrawArrays(GL_TRIANGLES, 0, 6);
-
-        // Now advance cursors for next glyph (note that advance is number of 1/64 pixels)
-        x += static_cast<float>(ch.Advance >> 6) * scale; // bitshift by 6 to get value in pixels (2^6 = 64)
-    }
-
-    glBindVertexArray(0);
-    glBindTexture(GL_TEXTURE_2D, 0);
-
-    glDisable(GL_CULL_FACE);
-    glDisable(GL_BLEND);
 }
